+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 
 - \_LEN / \_SIZE constans for null terminated strings
 - mDNS resolver example in `README.md`
+- Add dependencies
 
 ### Changed